requirement.txt
pillow
opencv-python
keras
fastapi
uvicorn
scikit-learn
werkzeug
tensorflow
python-multipart